build.sh: made some changes to script

+++ build.sh
@@ -20,19 +20,26 @@ echo
 echo "Compiling the source"
 echo
 PATH="/home/ktk/kernel/toolchain/custom-clang/bin:/home/ktk/kernel/toolchain/gcc9/bin:${PATH}"
-make -j2 O=out ARCH=arm64 CC=clang CLANG_TRIPLE=aarch64-linux-gnu- CROSS_COMPILE=aarch64-linux-gnu- CROSS_COMPILE_ARM32=~/kernel/toolchain/gcc-32/bin/arm-linux-androideabi-
-echo
-echo "The image is compiled"
-echo
-mv out/arch/arm64/boot/Image.gz-dtb ~/kernel/anykernel2/Image.gz-dtb
-rm -rf out
-cd ~/kernel/anykernel2
-echo
-echo "Creating a Flashable ZIP"
-echo
-zip -r9 noodle_kernel-$(date +"%Y-%m-%d").zip *
-rm -rf Image.gz-dtb
-echo
-echo "Movied to ZIPs Folder "
-echo
-mv noodle_kernel-$(date +"%Y-%m-%d").zip ~/kernel/ZIPs
+make -j4 O=out ARCH=arm64 CC=clang CLANG_TRIPLE=aarch64-linux-gnu- CROSS_COMPILE=aarch64-linux-gnu- CROSS_COMPILE_ARM32=~/kernel/toolchain/gcc-32/bin/arm-linux-androideabi-
+if [ -f out/arch/arm64/boot/Image.gz-dtb ]; then
+      echo
+      echo "The image is compiled"
+      echo
+      mv out/arch/arm64/boot/Image.gz-dtb ~/kernel/anykernel2/Image.gz-dtb
+      rm -rf out
+      cd ~/kernel/anykernel2
+      echo
+      echo "Creating a Flashable ZIP"
+      echo
+      zip -r9 noodle_kernel-$(date +"%Y-%m-%d").zip *
+      rm -rf Image.gz-dtb
+      echo
+      echo "Movied to ZIPs Folder "
+      echo
+      mv noodle_kernel-$(date +"%Y-%m-%d").zip ~/kernel/ZIPs
+else
+      echo
+  	  echo "Kernel did not compile, please check for errors!!"
+  	  echo
+fi
+rm -rf /out
